requirements.txt
Django~=4.2.5
django-ninja~=0.22.2